+++ snakemake/jobs.py
@@ -502,7 +502,7 @@ class Job(AbstractJob):
             raise RuleException(str(ex), rule=self.rule)
         except KeyError as ex:
             raise RuleException(
-                "Unknown variable in message " "of shell command: {}".format(str(ex)),
+                "Unknown variable in message of shell command: {}".format(str(ex)),
                 rule=self.rule,
             )
 
@@ -519,7 +519,7 @@ class Job(AbstractJob):
             raise RuleException(str(ex), rule=self.rule)
         except KeyError as ex:
             raise RuleException(
-                "Unknown variable when printing " "shell command: {}".format(str(ex)),
+                "Unknown variable when printing shell command: {}".format(str(ex)),
                 rule=self.rule,
             )